Add tests/unit/query_graph/sql/ast/test_datetime_diff.py
