+++ tests/fixtures/query_graph/test_datetime_diff/timestamp_tuple_schema_bigquery.sql
@@ -1,0 +1,7 @@
+(
+  DATEDIFF(
+    microsecond,
+    TO_TIMESTAMP(CAST(GET(zipped_column_2, 'timestamp') AS STRING), 'YYYY-MM-DD"T"HH24:MI:SS"Z"'),
+    TO_TIMESTAMP(CAST(GET(zipped_column_1, 'timestamp') AS STRING), 'YYYY-MM-DD"T"HH24:MI:SS"Z"')
+  ) * CAST(1 AS INT64) / CAST(1000000 AS INT64)
+)
